fixing logging in workers WIP 2
