+++ web/apps/photos/src/utils/file/index.ts
@@ -53,8 +53,8 @@ import {
 import { FileTypeInfo } from "types/upload";
 
 import { default as ElectronAPIs } from "@ente/shared/electron";
+import { workerBridge } from "@ente/shared/worker/worker-bridge";
 import { t } from "i18next";
-import imageProcessor from "services/imageProcessor";
 import { getFileExportPath, getUniqueFileExportName } from "utils/export";
 
 const WAIT_TIME_IMAGE_CONVERSION = 30 * 1000;
@@ -452,8 +452,7 @@ export async function getRenderableImage(fileName: string, imageBlob: Blob) {
                         imageBlob.size,
                     )}`,
                 );
-                throw new Error("bypass");
-                convertedImageBlob = await imageProcessor.convertToJPEG(
+                convertedImageBlob = await convertToJPEGInElectron(
                     imageBlob,
                     fileName,
                 );
@@ -484,6 +483,36 @@ export async function getRenderableImage(fileName: string, imageBlob: Blob) {
         return null;
     }
 }
+
+const convertToJPEGInElectron = async (
+    fileBlob: Blob,
+    filename: string,
+): Promise<Blob> => {
+    try {
+        const startTime = Date.now();
+        const inputFileData = new Uint8Array(await fileBlob.arrayBuffer());
+        const convertedFileData = await workerBridge.convertToJPEG(
+            inputFileData,
+            filename,
+        );
+        addLogLine(
+            `originalFileSize:${convertBytesToHumanReadable(
+                fileBlob?.size,
+            )},convertedFileSize:${convertBytesToHumanReadable(
+                convertedFileData?.length,
+            )},  native conversion time: ${Date.now() - startTime}ms `,
+        );
+        return new Blob([convertedFileData]);
+    } catch (e) {
+        if (
+            e.message !==
+            CustomError.WINDOWS_NATIVE_IMAGE_PROCESSING_NOT_SUPPORTED
+        ) {
+            logError(e, "failed to convert to jpeg natively");
+        }
+        throw e;
+    }
+};
 
 export function isFileHEIC(exactType: string) {
     return (
